tracks/track-04-agent-comms-slack-email/requirements.txt
# Core dependencies for Agent Communication Integration
# Participants will add platform-specific libraries as needed

# Standard Agent framework (install from GitHub)
# git+https://github.com/jentic/standard-agent.git

# Environment and configuration
python-dotenv>=1.0.0

# HTTP requests for API testing
requests>=2.31.0

# Basic async support
aiohttp>=3.9.0

# Logging and development
colorlog>=6.8.0

# Testing framework
pytest>=7.4.0

# Platform-specific libraries - UNCOMMENT and install as needed:

# For Slack integration:
slack-sdk>=3.26.0
slack-bolt>=1.18.0

# For Email integration (SendGrid):
# sendgrid>=6.10.0

# For Email integration (SMTP alternative):
secure-smtplib>=0.1.1

# For SMS integration (Twilio):
# twilio>=8.11.0

# Optional: Template processing for rich formatting
# jinja2>=3.1.0
# markdown>=3.5.0

# Optional: Additional communication platforms
# discord.py>=2.3.0  # Discord
# python-telegram-bot>=20.7  # Telegram